+++ SpaceShip.h
@@ -13,7 +13,7 @@ class SpaceShip : public GameObject,
 	public ISubscriber<XPEvent>
 {
 public:
-	SpaceShip(const char* name,TEXTURE_ID textureID, int initX, int initY,int width, int height, int order=0);
+	SpaceShip(const char* name,TEXTURE_ID textureID, float initX, float initY,int width, int height, int order=0);
 	void Update(float deltaTime) override;
 
 	void OnTriggerEvent(const XPEvent& eventType) override;
